+++ src/components/event-list.js
@@ -1,0 +1,10 @@
+const createEventListTemplate = () => {
+  return (
+    `
+    <ul class="trip-events__list">
+    </ul>
+    `
+  );
+};
+
+export {createEventListTemplate};
