+++ src/components/trip-item.js
@@ -1,0 +1,11 @@
+const createTripDayItemTemplate = () => {
+  return (
+    `
+    <li class="trip-days__item  day">
+      <div class="day__info"></div>
+    </li>
+    `
+  );
+};
+
+export {createTripDayItemTemplate};
