+++ src/components/trip-list.js
@@ -1,0 +1,10 @@
+const createTripDayWrapperTemplate = () => {
+  return (
+    `
+    <ul class="trip-days">
+    </ul>
+    `
+  );
+};
+
+export {createTripDayWrapperTemplate};
